Add volkInitializeCustom that takes a pointer to GIPA
In some cases, volkInitialize isn't sufficient to initialize Vulkan
loader since it assumes a specific .so path and a specific function to
load symbols from it.

Once we know the pointer to vkGetInstanceProcAddr, we can initialize all
other pointers; in cases where volkInitialize doesn't work it should
thus be sufficient to retrieve a pointer to GIPA and call
volkInitializeCustom which will remember the pointer to GIPA and
initialze all other function pointers using that.

Fixes #3.