Update to 1.1.86
Since 1.1.86 SDK adds a not-quite-ready extension that is marked as
disabled (VK_EXT_image_drm_format_modifier), this change extends
generate.py to exclude disabled extensions from the list - which leads
to VK_ANDROID_native_buffer being excluded as well.

This shouldn't be an issue since as far as I can tell, this extension
isn't present in any official Vulkan SDK headers so it's unlikely that
it could ever be used with volk.

No new entrypoints were added in 1.1.86 otherwise.