Remove dependencies on windows.h
This change makes sure that including volk.h doesn't pull in windows.h
with it - windows.h is a *very* heavy dependency, and it's not that
necessary since we only need a few symbols from it in the header.

As a result, including volk.h is now 200-100ms faster on a very fast CPU
with a hot FS cache depending on whether WIN32_LEAN_AND_MEAN is defined
or not.

Fixes #21.