Avoid re-loading same global entrypoints
Add static VkInstance and VkDevice vars to track the last
instance and device for which global entrypoints have been loaded.

When calling volkLoadInstance() or volkLoadDevice(), avoid
re-loading entrypoints for the same instance or device respectively.

This is useful for ANGLE where we iterate testing over different
single VkDevice objects on the same physical box. We can re-call the
volkLoad* functions and only incur the updates when needed.