Give precedence to libvulkan.so.1 over libvulkan.so (#43)
The correct dlopen behaviour is to load libraries by SONAME. The current
behaviour can cause different versions of libvulkan to be loaded at
runtime. And having multiple loaders in the same app can have undesired
side-effects.

We keep the libvulkan.so fallback for cases Android platforms where only
that name may be available.