Remove VK_HUAWEI_subpass_shading
This extension was introduced in 1.2.182, but then was changed in
1.2.184 (vkGetSubpassShadingMaxWorkgroupSizeHUAWEI was renamed to
vkGetDeviceSubpassShadingMaxWorkgroupSizeHUAWEI).

This change makes two versions of the SDK incompatible, since the
function pointer type name can't be used without knowing the version.

Thus this change removes the extension support outright until a time
when someone actually needs it; 1.2.182 checkpoint is also added to CI
build so that we don't only test master.