sdk: support the Vulkan SDK build process
The Vulkan SDK build process needs to be able to pass a CMake variable
in to the Volk build indicating where the Vulkan-Headers install build
resides.  Otherwise, Volk could end up building with an older version
of the SDK that happens to be installed on the build system (as our CI
systems have various different SDK/loader combinations installed, for
test purposes).

This change causes Volk to respect a VULKAN_HEADERS_INSTALL_DIR
CMake variable, if defined.  The variable must be checked before
the presence of a Vulkan SDK is checked, to ensure that a random
SDK installed on the system doesn't interfere with the build.