CMake: Fix include path discovery when using FindVulkan.cmake
volk does not need libvulkan.so dependency, and in fact adding it is
dangerous due to possible symbol conflicts and unnecessary as the point
of using volk is to avoid direct dependencies on Vulkan loader.

When FindVulkan.cmake discovered a Vulkan installation, it would
erroneously depend on Vulkan::Vulkan which would add libvulkan.so
dependency.

Instead, rework all of this to just get the include path from the module
or compute it from other variables. This setup is cleaner and it
guarantees that we aren't going to get any rogue flags into the build.