Fix conflict between FARPROC definition and minwindef.h
minwindef.h defines FARPROC without (void) argument list, so if it is
already included when volk.c is compiled via VOLK_IMPLEMENTATION, gcc
will emit an error (MSVC handles this mismatch fine).

Unfortunately, simply removing (void) results in a warning with
Wstrict-prototypes even when volk.c is compiled as standalone. We could
disable the warning with a pragma but for now we could simply reuse
FARPROC definition from minwindef.h if it's already included.