Implement instance table API
Similarly to having an option to bypass global instance function
pointers in favor of an instance function table, we now provide an API
to load instance pointers into a separate structure.

This makes it possible to create multiple instances and use them
concurrently; this isn't needed as often as multi-device API but can
still be valuable for some applications.