Update volkFinalize header comments to clarify usage
volkFinalize is only useful when volkInitialize succeeded and the caller
wants to unload vulkan-1.dll/.so - for example, maybe when switching
between Vulkan and OpenGL renderers an application might want a cleaner
set of loaded modules. This addresses a fairly rare use case, so we now
clarify that the usage is optional.

Fixes #156.