Fall back to /usr/local/lib/libvulkan.dylib on macOS
dlopen on macOS is supposed to automatically search /usr/local/lib per dlopen man page.
This behavior is customizable via DYLD_LIBRARY_PATH and DYLD_FALLBACK_LIBRARY_PATH, but
even if neither is explicitly set /usr/local/lib should be a search path. Unfortunately,
this does not happen anymore on modern macOS systems, and this is precisely where
Vulkan SDK installs dylib by default.

This used to "work" in CMake builds that accidentally linked in libvulkan.dylib statically,
because the library was already loaded into the process, but fixing that exposed this issue.

For now we need to manually fallback to /usr/local/lib path.