Remove dependency on Xlib.h
When building volk with VK_USE_PLATFORM_XLIB_KHR, volk includes Xlib.h
as a byproduct of including vulkan.h.

Xlib.h is not a very nice header: it defines a lot of unprefixed macros
that may interfere with other code included after volk.h (e.g. it
defines Convex which breaks Jolt Physics headers).

On Windows, we've been using a change that includes vulkan_win32.h
manually after redefining the subset of windows.h that it needs. This
change attempts to do the same for Xlib.h.

Unfortunately, this requires duplicating pretty much the entire vulkan.h
to replace these bits; for Windows, there's only one expected platform
that can be active at a given time, but on Linux applications might want
up to 4 in rare cases (Xlib, Xlib Randr, Xcb, Directfb). The alternative
would be to #undef the relevant platform defines before including
vulkan.h, but they need to be redefined after include, so it's not clear
that this is much simpler.

Crucially, we define the necessary types exactly like that's done in
Xlib.h unconditionally; this ensures that any type mismatches do not
depend on the header inclusion order and are caught early.