Generate padding fields in VolkDeviceTable to preserve ABI
To avoid issues with ABI compatibility of volk built with different sets
of defines, we now generate padding fields for each extension block with
the same number of function pointers. This should make sure that the
offset of every pointer for any given volk version remains stable,
simplifying binary builds.