Use include path from Vulkan-Headers target if present
When Vulkan-Headers is used by the application, we can use the include path
from it even though other ways to discover the include paths fail. This
simplifies the usage of volk via CPM and when using it directly; it does
require the application to include Vulkan-Headers first, as there's no way
to discover it otherwise.